Add tests
The pytest library seems to do something weird with the device, making
any batches of GPU launches after the first one to always fail... You
can confirm it is not the function itself failing by either running the
same parameters in torchmd-net/neighbors/demo.py or switching the
order of the pytest parameters to appear at the beginning.

For example, in this commit the cases for radius equal to 10 fail. If
you change

-@pytest.mark.parametrize('radius', [8, 10])
+@pytest.mark.parametrize('radius', [10, 8])

suddenly the cases for radius 8 are the ones to fail, while the cases
for radius 10 work without a problem.

There is no problem with the CPU executions.